Fixed code style

+++ pavel_aleynikov/src/test/java/task/CountDNANucleotidesCalculatorTest.java
@@ -10,15 +10,15 @@ public class CountDNANucleotidesCalculatorTest {
     @Test
     void testOnNormalDataset() {
         final String dna = "AGCTTTTCATTCTGACTGCAACGGGCAATATGTCTCTGTGTGGATTAAAAAAAGAGTGTCTGATAGCAGC";
-        Map<Character, Long> result = CountDNANucleotidesCalculator.calculate(dna);
-        Assertions.assertEquals(Map.of('A', 20L, 'C', 12L, 'G', 17L, 'T', 21L), result);
+        Map<Character, Integer> result = CountDNANucleotidesCalculator.calculate(dna);
+        Assertions.assertEquals(Map.of('A', 20, 'C', 12, 'G', 17, 'T', 21), result);
     }
 
     @Test
     void testOnEmptyDataset() {
         final String dna = "";
-        Map<Character, Long> result = CountDNANucleotidesCalculator.calculate(dna);
-        Assertions.assertEquals(Map.of('A', 0L, 'C', 0L, 'G', 0L, 'T', 0L), result);
+        Map<Character, Integer> result = CountDNANucleotidesCalculator.calculate(dna);
+        Assertions.assertEquals(Map.of('A', 0, 'C', 0, 'G', 0, 'T', 0), result);
     }
 
 }
